feat_&_fix: major update 3.0
1) ProjectMamager:
fixed artifacts in the project image icon

2) Project Settings
fixed a bug where you had to change the project name to save it

3) JSON Saving
fixed a bug where saving the same project duplicated records about its images

4) Sound
fixed a bug where the sound disappeared when the project was reopened.

5) Metronome
5.1) fixed the 1/16 forward offset
5.2) it is now synchronized with the timer constantly. To turn it off, simply mute it

6) track:
6.1) fix setting of 16ths measures of tackt
6.2) now after resetting the audio file the track stops
6.3) now after turn off the loop the track stops
6.4) implemented lag and duration of the track

7) TrackSettings
7.1) Removed the getLoopState method
7.2) the record button type was replaced with RedButton instead of QCheckBox
7.3) removed setters for lag and duration with a size of 1/16 of a bar
7.4) redesigned the layout of the manipulators

8) Red button:
8.1) RedButton now after width
8.2) names of some methods changed

9) LCD display
removed FixedSize

10) LCDCounter
10.1) now it is not a QSpinBox. Now it derriving QWidget
10.2) reworked buttons

+++ ProjectView.h
@@ -24,8 +24,8 @@ private:
 
 public:
     explicit ProjectView(
-        quint16 width,
-        quint16 project_bpm,
+        int width,
+        int project_bpm,
         qint64 seconds_of_last_use,
         const QString& project_name,
         const QString& path_to_project,
